Add files via upload
This report examines graduate school acceptance probability as a function of several predictor variables such as College GPA, Graduate Record Examination, TOEFL, Statement of Purpose, and Letter Recommendation. The aim of this report is to build a linear regression model which can predict the graduate school acceptance probability with a suitable amount of precision to understand to the extent of each predictor’s relation to graduate school acceptance probability.