Add files via upload
The dataset used in this model came from the game FIFA 19. It contained various attributes about each player and their information stored within the game. The problem at hand was to see if a multiple linear regression model could be used to predict a player’s Overall rating in the game. The attributes chosen were all continuous which we thought were crucial in determining a player’s rating within the game. The model found that certain predictors and not all were very significant in predicting a player’s Overall Rating. Absence of one variable did not affect other predictors as much because some variables were already insignificant.