Add files via upload
What makes fine wine? The red wine variant of the Portuguese “Vinho Verde” wine refers to Portuguese wine that originated in the historic Minho province in the far north of the country. The main goal of this investigation is to find which features of these kinds of wine provide the most information about its quality using logistical regression.